Walking and Jumping fixed
You can see everything I'm doing within the Julian folder. I'm making copies of the levels where I am testing and changing stuff.

So far I've done the basic walking and jumping script for the player. It works using layer masks, so if you want to use the player, make sure to apply the "Ground" layer on objects where the player can jump.

That's it for now :)